refactor drop-down
